requirements.txt
autopep8==1.4.4
gitdb2==2.0.6
GitPython==3.0.4
pycodestyle==2.5.0
smmap2==2.0.5
